Added .sql to this repository.
We're moving more towards using SQL instead of using Python to move the
data from the KanRoad system to the KanDrive system due to some
performance issues. This revision includes the code that was used to
test and develop the new processes to move the data. It also includes
the code that resulted from that testing and development.